# Finer graph locks