@MHBauer 2016-02-12
I like that you added a test/benchmark before going at it.