@calavera 2016-02-12
@cpuguy83 are you going to do the same for the rest of the drivers? I think all of them have the same problem.