@cpuguy83 2016-02-12
@calavera That was my plan, though I have to admit for some reason now I'm not getting those numbers I posted above (which I originally ran multiple times), so looking into it.