@GordonTheTurtle 2016-02-13
Job: Docker-PRs-WoW-TP4 [FAILED](https://jenkins.dockerproject.org/job/Docker-PRs-WoW-TP4/1010/console):

~~~console

---
	      56	docker-1.11.0-dev.exe.md5  0%  100%  
	    New File  		      88	docker-1.11.0-dev.exe.sha256  0%  100%  
	    New File  		  27.1 m	docker.exe  0.0%  3.6%  7.3% 11.0% 14.7% 18.3% 22.0% 25.7% 29.4% 33.0% 36.7% 40.4% 44.1% 47.7% 51.4% 55.1% 58.8% 62.5% 66.1% 69.8% 73.5% 77.2% 80.8% 84.5% 88.2% 91.9% 95.5% 99.2%100%  

------------------------------------------------------------------------------

               Total    Copied   Skipped  Mismatch    FAILED    Extras
    Dirs :         1         0         1         0         0         0
   Files :         4         4         0         0         0         0
   Bytes :   54.39 m   54.39 m         0         0         0         0
   Times :   0:00:00   0:00:00                       0:00:00   0:00:00


   Speed :           1188289583 Bytes/sec.
   Speed :            67994.475 MegaBytes/min.
   Ended : Saturday, February 13, 2016 12:57:04 AM

+ ec=0
+ set +x
INFO: Linking the built binary to /d/CI/CI-7a
---

~~~