@sillyousu 2016-02-15
Maybe another issue: I think the `Clean` method need a read lock. 
https://github.com/cpuguy83/docker/blob/finer_graph_locks/daemon/graphdriver/aufs/aufs.go#L486