@GordonTheTurtle 2016-02-17
Job: Docker-PRs-WoW-TP4 [FAILED](https://jenkins.dockerproject.org/job/Docker-PRs-WoW-TP4/1082/console):

~~~console

---
	      56	docker-1.11.0-dev.exe.md5  0%  100%  
	    New File  		      88	docker-1.11.0-dev.exe.sha256  0%  100%  
	    New File  		  27.1 m	docker.exe  0.0%  3.6%  7.3% 11.0% 14.7% 18.4% 22.1% 25.7% 29.4% 33.1% 36.8% 40.5% 44.2% 47.9% 51.5% 55.2% 58.9% 62.6% 66.3% 70.0% 73.7% 77.3% 81.0% 84.7% 88.4% 92.1% 95.8% 99.5%100%  

------------------------------------------------------------------------------

               Total    Copied   Skipped  Mismatch    FAILED    Extras
    Dirs :         1         0         1         0         0         0
   Files :         4         4         0         0         0         0
   Bytes :   54.27 m   54.27 m         0         0         0         0
   Times :   0:00:00   0:00:00                       0:00:00   0:00:00


   Speed :           1138148240 Bytes/sec.
   Speed :            65125.364 MegaBytes/min.
   Ended : Wednesday, February 17, 2016 2:39:33 AM

+ ec=0
+ set +x
INFO: Linking the built binary to /d/CI/CI-72
---

~~~