@cpuguy83 2016-02-17
Thanks @sillyousu

Although this probably won't ever be called but by a single goroutine and only while it has exclussive access to the graphdriver, I agree it should have a lock.

I've added one, but I'm not sure it is the best solution since it is technically feasible for other operations to be mounting things while `Cleanup` is trying to unmount... but again I don't think the engine can even call things these in parallel.