@cpuguy83 2016-02-25
Ok, I've moved this into two commits. 1 for fixing some of the issues with aufs, the other to add the id-based locking.