@LK4D4 2016-02-25
@cpuguy83 it's ok for me. but don't you think that per-object lock will solve issue in more performant way(no allocations on first lock etc.)?