@MHBauer 2016-02-25
I kinda liked the commit with the bench only. I can't get it to actually run on the first commit.

Is the `auplink` error message an actual error? I'm getting that running the bench.

Here's what I use to get results on the second commit:
```
λ TESTFLAGS='-bench=. -cpu 2,4,8 -benchtime 10s' TESTDIRS='daemon/graphdriver/aufs/' make test-unit
```
results
``` 
BenchmarkConcurrentAccess  	time="2016-02-25T21:48:38Z" level=error msg="Couldn't run auplink before unmount /tmp/aufs-tests/aufs/mnt/c775c928309c3b7f49caecc92323a21508b707e6390d775812e5b356b8e49d53: exit status 22"
      10	1424157163 ns/op
BenchmarkConcurrentAccess-2	      20	 812425911 ns/op
BenchmarkConcurrentAccess-4	      10	1127573648 ns/op
BenchmarkConcurrentAccess-8	      10	1457461355 ns/op
coverage: 80.1% of statements
ok  	github.com/docker/docker/daemon/graphdriver/aufs	114.734s
```