@cpuguy83 2016-02-25
Updated per @LK4D4's suggestion and posted new bench results.