@MHBauer 2016-02-26
@cpuguy83 It seems to hang. Maybe it's just taking a long time.