@cpuguy83 2016-02-26
@MHBauer It does create a lot of contention, but still finishes pretty fast on mine system... are you on an SSD?