@MHBauer 2016-02-26
ssd but in a vm. OSX-> docker-machine/virtualbox. Maybe I should set up a dev-vm someday.

I'm certainly not saying anything is wrong. I want to run it. I am curious.