@cpuguy83 2016-02-26
@MHBauer I'd prefer to make sure there is nothing wrong, also you have my same setup :)