@MHBauer 2016-02-26
Well, just to be clear, the f310141 commit works great. No problems with that. Tests run and finish.

I'm only talking about 55c91f2. It compiles and then `aufs.test` goes to sleep.